[FEAT] updated pi0&smolVLA with the same parent class
